20250407_update(1)

+++ teamjo/src/main/java/com/smhrd/teamjo/entity/UserInfo.java
@@ -60,5 +60,5 @@ public class UserInfo {
     private String nick;
 
     @Column(name = "RECOM_CAL")
-    private BigDecimal recomCal;
+    private Integer recomCal;
 }
